yaprompt_python/requirements.txt
pydantic>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
